requirements.txt
feedparser==6.0.11
jsonpickle==3.0.2
PyYAML==6.0.1
sgmllib3k==1.0.0
